refactor(sof): expose the real server router from the lib
Integration tests now build the production router instead of a duplicated
stub. Stub-drifted assertions were corrected to the real contract; the
unreachable 422 case is quarantined and tracked in #670.

Fixes #576
